refactor tests a little
